examples/mcp/filesystem_example/requirements.txt
agents
python-dotenv>=1.0.0
openai>=1.12.0
nest-asyncio>=1.5.8
requests>=2.31.0